add_shell: get __builtins__ directly instead of from __main__
In some situations, __main__.__builtins__ doesn't exist at the time
add_shell() is called.  The old code assumed that it does, and would segfault.

Before this commit:
$ python -c 'import IPython; \
             app = IPython.terminal.ipapp.TerminalIPythonApp.instance(); \
             app.user_ns = {}; app.initialize(); app.start()'
In [1]: import apsw
<segmentation fault>

This was first discovered when using pyflyby:
$ py
In [1]: import apsw
<segmentation fault>

To avoid that problem, this commit changes the code to import __builtin__ as a
module directly.