Only use multiprocessing with fork
Fixes #277

We use the multiprocessing module for testing the fork checker, as
this is the most likely (unintentional) way a program would end up
forking with an open database.

multiprocessing used to use fork on all platforms that had it, but
has recently changed to spawn on MacOS.  So only use multiprocessing
when it is forking.