Failed attempt to simplify the code, reverted in next commit
This is committed for reference and comedy,  The fundamental
problem is that line break rules can't be applied in order
because there are later rules that match longer sequences
than earlier rules, and the later ones are expected to win.

That leads to a mess, which is what I started with before
the simplification.  The mess passed all the tests, except
a small number around LB25