Remove the concept of html non-spacing tags
The idea was that these tags could occur within a word
like h<b>ell</b>o but this resulted in the close tag being
considered part of the word, so one</span> would include
the </span> as part of the utf8 end offset.  That leads to ugly
snippet/highlight etc.

We now consider all tags as introducing space therefore providing
a word boundary since that is the most common in HTML I've examined.