Update function parameter types to always be PyObject *
Standard Python dev practice has been to use the "subclass" C type
for the self parameter.  The methods are called with PyObject * as the
type.  This does not cause any problems.

Except control flow integrity is now available.  It requires that
the types and parameters of a called function match exactly, and if
not will fail.  clangs undefined sanitizer warns about these.

CPython fixed their own calls in 3.14 - see
https://github.com/python/cpython/issues/111178

This large change corrects the self parameter to PyObject * and
then casts to the useful type in the body.  It also adds the various
unused closure parameters.  It won't cause any code generation changes.

Refs #498