Fix cursor inheritance of tracers
You can set tracers on the connection and the cursor.  If not set
on the cursor, it defaults to whatever is set on the connection.

But there is one more case where it is set on the connection, and
you do not want a tracer on the cursor.  The semantics I intended
is that setting it to None on the cursor would then *not* inherit
from the connection.  What got implemented was that setting to
None turned inheritance back on.

This commit fixes that.

Over all this time nobody noticed the issue so I won't document
this fix.