Delete AnyIO
It has the same problem as asyncio which is that sending
the coroutine to the event loop with a timeout and
having that timeout expire does not return that
timeouterror back to the worker thread and its call waiting
on the result.

With asyncio we have a future and have to supply a timeout
to that too.  With Trio it works correctly.

Realistically nobody will use the AnyIO wrapper since one
of the other two is below the hood.